requirements.txt
# Text Processing
nltk==3.8.1
# Ensure `punkt` tokenizer is available for NLTK
# Run `nltk.download('punkt')` after installation

# PDF Parsing
pdfplumber==0.10.1
pytesseract==0.3.10  # Only if OCR is needed for scanned PDFs

# Sentence Embeddings
sentence-transformers==2.2.2
sentencepiece==0.2.0
transformers==4.47.0

# Vector Store
faiss-cpu==1.7.4  # Use faiss-gpu if you have a compatible GPU

# Question Answering
transformers==4.35.0
torch>=2.0.0  # Required for Transformers and SentenceTransformer

# Environment Management
python-dotenv==1.0.0  # For loading environment variables

# Streamlit for Frontend
streamlit==1.25.0

# Additional Utilities
numpy>=1.26.0
pandas==2.1.2  # Optional for structured data handling
scikit-learn==1.4.2  # For additional ML utilities, if needed

# LangChain (specific modules referenced in imports)
langchain-core==0.3.29  # Update as per your compatibility needs
langchain-community==0.3.14  # Includes Ollama
langchain-google-genai>=2.0.0

# Testing
pytest==7.4.2

# Optional OCR Tools (only if OCR support is added)
pillow<10s
